backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1 
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0